refactor: modularize websocket and UI logic
